add rust setup

+++ PowerShell/Profile.ps1
@@ -4,7 +4,9 @@ Function d {Set-Location "~\dotfiles\"}
 Function .. {Set-Location "..\$args"}
 Function ... {Set-Location "..\..\$args"}
 Function .... {Set-Location "..\..\..\$args"}
-Remove-Item alias:gp -Force
+if( (Get-Alias gp).CommandType -eq "Alias" ){
+  Remove-Item alias:gp -Force
+}
 Function gp {git pull}
 Set-Alias -Name which -Value where.exe
 Set-Alias -Name vi -Value nvim
